bin: Add small shell scritp that kill processes
Cold blooded process killer. He will track them down and kill them.
Let's put the jokes away. The script expects the name of a poress
as an argument, if is not pressent it will ask you kindly for a one
"I need a name boss!" . If  an agrument is present but there is not
 such a process active at the moment it will notify you
"<PROCESS> is not around now! I can't whack him boss!".
And lastly if ther is a process and it is active , it will kill it.
The script is designed to kill processes that listen to some machine
ports such as node server, mongoDB and react server etc. Plays
very well with another script in this folder called ports, that displays
all processes that listen to some port.