Add more functionality to color script
Now we can get the full palette of hex values of a given number via
curl GET request. Now is easy to push them in one .css file and
pipe them in a vs code instance. Now we have the same functionality
of color palette check, but we don't need to go to the browser at
www.0to256.com.