vim: Refactor copule of bugs left in the .vimrc
First bug was in vim-jsx somewhere. I couldn't fix it, so I decide to
manually force vim to change filetype to javascript.jsx. This is the
only filetype that supports jsx and javascript in the same time.

The second one was problem in the indentation of brackets in javascript.
I have played a bit with all indent related settings, so I have done it
and works now. All related setting now lie in dedicated section in my
.vimrc.