[fix] Fix git_info function in dnonov.zsh-theme
Getting a meaningful branch output was possible only if you checkout
different branches and not detaching the HEAD. However if you detach
the HEAD the former algorithm was not sufficient to give you the right
output '(HEAD detached at 34ff32a)', instead it was '(HEAD'.

One of the problems was how to find the current branch. So the former
implementation was a function that executes 'git branch', the output of
which is all the branches smashed into a line. Because of the specifics
of 'git branch' output there is a caret in front of the current branch
so if you cut on that caret and get the second field, pipe it down to
another cut with a delimiter white space, the second field must be
the current branch and it is, only if your HEAD is not detached.

The straight forward way of doing all this is just to grep for the caret.
Then cut on white space and get the second field.