[fix] Remove ~/.eslintrc.json
The tsserver in a typscript project was picking my ~/.eslintrc.json.
So linting was a nightmare.

This was possible because of the tsserver config search mechanics.
First it would look into your project dir and if can't find a .eslintrc
file it would go one dir up and check there, and so on and so forth all
the way up to your home dir.

Another problem was I had only coc-eslint installed, if I had coc-tslint
as well, tsserver it'd pick the local .tslintrc file and wouldn't went
all the way up to my home dir.

The solution is just to keep my home dir empty of any linting configs.
And add a script to generate a local .eslintrc.json file if I want to.