[feat] Enable virtual text and diagnostic in nvim
It is very important to keep all diagnostic signs as they are.
Only diagnostic.infoSign is 3 characters long, because I'm using a
corner case in COC that if you pass a string longer than 2 characters
it won't be displayed in the gutter. However if all of the signs are
more than 2 characters long all diagnostics are not displayed.

So I was able to solve a problem of my vim config. Clashing of
git-gutter and coc in the gutter. Coc always has priority over
git-gutter so if there is a spell info in the same line of a git change.
I would not see the git change in the gutter.

Now I get rid of the info signs in the gutter and replaced them with
codelens.